26-API: Rename `Okane.Api.Tests.Tests` to `Okane.Api.Tests.Testing`
The "Tests" folder doesn't contain any tests. Rather, it contains
configuration used in the actual tests.

"Testing" isn't the greatest name either, but it's less misleading than
"Tests".
